requirements.txt
os
json
camelot-py[cv]
PyPDF2==2.11.1
pandas